+++ NEWS.md
@@ -8,7 +8,10 @@ output: html_document
 
 ## Changes
 - `sp` moved to suggests, users are now prompted to install sp if they want to search stations by coordinates
-
+- `weather_dl()` replaces `weather()`
+- `weather_interp()` replaces `add_weather()`
+- `stations_dl()` replaces `stations_all()`
+- `tz_calc()` replaces `get_tz()`
 
 # weathercan 0.2.3 (2017-11-22)